+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/IndentationRule.kt
@@ -76,6 +76,7 @@ import com.pinterest.ktlint.core.ast.nextLeaf
 import com.pinterest.ktlint.core.ast.nextSibling
 import com.pinterest.ktlint.core.ast.parent
 import com.pinterest.ktlint.core.ast.prevCodeLeaf
+import com.pinterest.ktlint.core.ast.prevCodeSibling
 import com.pinterest.ktlint.core.ast.prevLeaf
 import com.pinterest.ktlint.core.ast.prevSibling
 import com.pinterest.ktlint.core.ast.upsertWhitespaceAfterMe
@@ -569,6 +570,8 @@ class IndentationRule : Rule("indent"), Rule.Modifier.RestrictToRootLast {
                             ctx.ignored.add(n)
                         }
                     }
+                    FUNCTION_LITERAL ->
+                        adjustExpectedIndentInFunctionLiteral(n, ctx)
                     WHITE_SPACE ->
                         if (n.textContains('\n')) {
                             if (
@@ -791,6 +794,33 @@ class IndentationRule : Rule("indent"), Rule.Modifier.RestrictToRootLast {
         expectedIndent++
         debug { "++inside(CONDITION) -> $expectedIndent" }
         ctx.exitAdjBy(n.treeParent, -1)
+    }
+
+    private fun adjustExpectedIndentInFunctionLiteral(n: ASTNode, ctx: IndentContext) {
+        require(n.elementType == FUNCTION_LITERAL)
+
+        var countNonWhiteSpaceElementsBeforeArrow = 0
+        var arrowNode: ASTNode? = null
+        var hasWhiteSpaceWithNewLine = false
+        val iterator = n.children().iterator()
+        while (iterator.hasNext()) {
+            val child = iterator.next()
+            if (child.elementType == ARROW) {
+                arrowNode = child
+                break
+            }
+            if (child.elementType == WHITE_SPACE) {
+                hasWhiteSpaceWithNewLine = hasWhiteSpaceWithNewLine || child.text.contains("\n")
+            } else {
+                countNonWhiteSpaceElementsBeforeArrow++
+            }
+        }
+
+        if (arrowNode != null && hasWhiteSpaceWithNewLine) {
+            expectedIndent++
+            debug { "++after(FUNCTION_LITERAL) -> $expectedIndent" }
+            ctx.exitAdjBy(arrowNode.prevCodeSibling()!!, -1)
+        }
     }
 
     private fun indentStringTemplate(